+++ quality/refactor_health_report.md
@@ -552,17 +552,23 @@ Portfolio liquidity and projected-cashflow response assembly have been extracted
 689 script-counted lines while preserving upstream loading, warning, partial-failure, cash-balance,
 and cashflow response behavior. The blocking source-file threshold is ratcheted from 714 to 712
 lines, making `src/app/clients/advise_client.py` the largest residual source-file hotspot.
+Advise proposal lifecycle and memo upstream forwarding has been extracted into
+`src/app/clients/advise_proposal_client.py` while preserving the public `AdviseClient` surface.
+`src/app/clients/advise_client.py` is reduced from 712 to 220 script-counted lines,
+`src/app/clients/advise_proposal_client.py` is 536 script-counted lines, and the blocking
+source-file threshold is ratcheted from 712 to 709 lines, making
+`src/app/contracts/risk_workspace.py` the largest residual source-file hotspot.
 
 ## Health Signals
 
 | Area | Current posture | Evidence |
 | --- | --- | --- |
-| Branch hygiene | Healthy | Current portfolio liquidity response-boundary branch was created from clean `main`; no open PRs and no unmerged remote branches were present at slice start; final remote/local cleanup remains a post-merge gate |
-| Unit/contract coverage | Healthy | Focused validation passed with 46 portfolio service/liquidity tests, mypy over the touched service modules, and refactor-threshold trials proving `max_source_file_lines=712` passes while `711` fails on `src/app/clients/advise_client.py`; full local `make check` passed with 1,196 unit/contract tests |
+| Branch hygiene | Healthy | Current Advise proposal client-boundary branch was created from clean `main`; no open PRs and no unmerged remote branches were present at slice start; final remote/local cleanup remains a post-merge gate |
+| Unit/contract coverage | Healthy | Focused validation passed with 28 Advise client boundary/upstream tests, 4 refactor-threshold tests, mypy over the touched client modules, and refactor-threshold trials proving `max_source_file_lines=709` passes while `708` fails on `src/app/contracts/risk_workspace.py`; full local `make check` remains a pre-PR/merge gate for this branch |
 | Integration coverage | Healthy | Full local `make ci` passed with migration contract smoke and 209 integration tests |
 | Total coverage | Healthy | Full local `make ci` passed with 1,405 combined coverage tests and 94.27% total coverage |
-| Security audit | Governed | Current portfolio liquidity response-boundary branch introduces no dependency, authentication, caller-context, product-error-detail, or monetary-float policy changes; full local `make ci` security audit reported no known vulnerabilities after the governed `PYSEC-2026-161` exception |
-| Modularity | Improving, incomplete | Longest-function baseline remains 49 lines; recent response/request-context/parser/mapper/contract slices lower `portfolio_service.py` to 689 script-counted lines, `advisor_brief_source.py` to 508 script-counted lines, `risk_workspace_service.py` to 633 script-counted lines, `performance_workspace_evidence.py` to 461 script-counted lines, `workbench.py` to 47 script-counted lines, `performance_workspace_service.py` to 639 lines, `advisor_brief_service.py` to 435 script-counted lines, `proposal_service.py` to 658 script-counted lines, `dpm_command_center_service.py` to 695 lines, `dpm_wave_service.py` to 700 script-counted lines, `advise_client.py` to 712 script-counted lines, `dpm_client.py` to 452 script-counted lines, `src/app/contracts/dpm_command_center.py` to 593 script-counted lines, `src/app/contracts/proposals.py` to 431 script-counted lines, `src/app/contracts/advisor_brief.py` to 646 script-counted lines, `src/app/contracts/portfolio.py` to 281 script-counted lines, `src/app/contracts/portfolio_workspace.py` to 503 script-counted lines, `src/app/services/foundation_service.py` to 618 script-counted lines, `src/app/contracts/performance_workspace.py` to 651 script-counted lines, and `src/app/router_registry.py` to 632 script-counted lines; `src/app/clients/advise_client.py` is now the largest current source-file hotspot at 712 lines |
+| Security audit | Governed | Current Advise proposal client-boundary branch introduces no dependency, authentication, caller-context, product-error-detail, monetary-float policy, or upstream error-shape changes; full local `make ci` security audit remains a pre-PR/merge gate for this branch |
+| Modularity | Improving, incomplete | Longest-function baseline remains 49 lines; recent response/request-context/parser/mapper/contract slices lower `portfolio_service.py` to 689 script-counted lines, `advisor_brief_source.py` to 508 script-counted lines, `risk_workspace_service.py` to 633 script-counted lines, `performance_workspace_evidence.py` to 461 script-counted lines, `workbench.py` to 47 script-counted lines, `performance_workspace_service.py` to 639 lines, `advisor_brief_service.py` to 435 script-counted lines, `proposal_service.py` to 658 script-counted lines, `dpm_command_center_service.py` to 695 lines, `dpm_wave_service.py` to 700 script-counted lines, `advise_client.py` to 220 script-counted lines, `advise_proposal_client.py` to 536 script-counted lines, `dpm_client.py` to 452 script-counted lines, `src/app/contracts/dpm_command_center.py` to 593 script-counted lines, `src/app/contracts/proposals.py` to 431 script-counted lines, `src/app/contracts/advisor_brief.py` to 646 script-counted lines, `src/app/contracts/portfolio.py` to 281 script-counted lines, `src/app/contracts/portfolio_workspace.py` to 503 script-counted lines, `src/app/services/foundation_service.py` to 618 script-counted lines, `src/app/contracts/performance_workspace.py` to 651 script-counted lines, and `src/app/router_registry.py` to 632 script-counted lines; `src/app/contracts/risk_workspace.py` is now the largest current source-file hotspot at 709 lines |
 | API governance | Improving, incomplete | 233 OpenAPI paths and 247 operations have summaries, descriptions, operation IDs, tags, and documented 4xx/5xx responses; Spectral remains report-only |
 | Error consistency | Improving, incomplete | Reporting job and report-batch upstream error handling now uses explicit code-owned mapping rules with focused product-safe fallback tests; shared generic service-error status mapping is code-owned and tested; advisory-facing product-safe service-error defaults now use typed immutable configs; broader route/upstream error normalization remains open |
 | Architecture rules | Improving, incomplete | AST boundary tests exist; import-linter is report-only; source-file and function-size thresholds are now blocking through `make lint` |
@@ -570,7 +576,7 @@ lines, making `src/app/clients/advise_client.py` the largest residual source-fil
 
 ## Primary Refactor Backlog
 
-1. Continue splitting large contract modules, `advise_client.py`, and remaining upstream
+1. Continue splitting large contract modules and remaining upstream
    clients around clear service and route-family boundaries.
 2. Continue splitting `portfolio_service.py` around clear source-readiness, workspace, insight,
    book, and workflow adapters only when cohesive behavior-preserving seams remain.
@@ -609,7 +615,7 @@ lines, making `src/app/clients/advise_client.py` the largest residual source-fil
 
 1. Report-only workflow uploads quality logs for baseline classification.
 2. Blocking refactor threshold gate now enforces:
-   - no Python source file under `src/app` above 712 script-counted lines,
+   - no Python source file under `src/app` above 709 script-counted lines,
    - no Python function or async function above the remediated 49-line AST span baseline.
 3. Then enforce no-new-regression thresholds for:
    - ruff/mypy,
