+++ wiki/Architecture.md
@@ -10,11 +10,11 @@
 - consolidated architecture and quality-baseline docs under
   [docs/architecture.md](../docs/architecture.md) and
   [quality/architecture_rules.md](../quality/architecture_rules.md)
-- current enterprise-hardening evidence records `src/app/clients/advise_client.py` as the
-  largest residual source file at 712 script-counted lines after portfolio liquidity and
-  projected-cashflow response assembly moved into a focused helper module; upstream loading remains
-  in `PortfolioService`, and the repository longest-function baseline remains held at the enforced
-  49-line AST span
+- current enterprise-hardening evidence records `src/app/contracts/risk_workspace.py` as the
+  largest residual source file at 709 script-counted lines after proposal lifecycle and memo
+  forwarding moved into `src/app/clients/advise_proposal_client.py`; the public `AdviseClient`
+  surface remains preserved, and the repository longest-function baseline remains held at the
+  enforced 49-line AST span
 
 ## Route-family map
