+++ wiki/Validation-and-CI.md
@@ -646,6 +646,17 @@ refactor-threshold trials proving `max_source_file_lines=712` passes while `711`
 migration contract smoke, 209 integration tests, 1,405 combined coverage tests, 94.27% total
 coverage, and no known vulnerabilities after the governed `PYSEC-2026-161` exception.
 
+The current Advise proposal client-boundary branch moves proposal lifecycle and memo upstream
+forwarding into `src/app/clients/advise_proposal_client.py` while preserving the public
+`AdviseClient` surface. It reduces `advise_client.py` from 712 to 220 script-counted lines, keeps
+the new proposal mixin at 536 script-counted lines, moves the largest residual source-file hotspot
+to `src/app/contracts/risk_workspace.py` at 709 script-counted lines, and ratchets the source-file
+threshold to 709 script-counted lines. Focused validation passed with ruff check, ruff format
+check, mypy over touched client modules, 28 focused Advise client boundary/upstream tests,
+4 refactor-threshold tests, and refactor-threshold trials proving `max_source_file_lines=709`
+passes while `708` fails on `src/app/contracts/risk_workspace.py`. Full local `make check` and
+`make ci` remain the pre-PR/merge gates for this branch.
+
 The current portfolio transaction-summary context branch moves reporting-window resolution, YTD
 transaction pagination, defensive page-row extraction, reporting-currency fallback, and
 requested-window filtering into `portfolio_transaction_summary.py`. It reduces
